% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simulations.R
\name{plot_simulations}
\alias{plot_simulations}
\title{Plot the results of a simulation run}
\usage{
plot_simulations(results, xlab = "Simulation", ylab = "Result",
  ylim = NULL, true_value = NULL, error = F, sort = T,
  main = NULL)
}
\arguments{
\item{results}{A vector specifying the value, lower and higher limit of the corresponding confidence interval.}

\item{xlab}{Lorem ipsum.}

\item{ylab}{Lorem ipsum.}

\item{ylim}{Lorem ipsum.}

\item{true_value}{Lorem ipsum.}

\item{error}{Lorem ipsum.}

\item{sort}{Lorem ipsum.}

\item{main}{Lorem ipsum.}
}
\value{
Plot
}
\description{
Lorem ipsum
}
\examples{
# Create a default universe with 50 ng DNA input.
universe(50)
}
