% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_universe.R
\name{sample_from_universe}
\alias{sample_from_universe}
\title{Take a sample from an existing universe.}
\usage{
sample_from_universe(universe, sample, alpha = 0.95)
}
\arguments{
\item{universe}{An existing universe.}

\item{sample}{An integer specifying the size of the sample to be taken, or a vector indicating which elements of the universe should be sampled.}

\item{alpha}{Level of significance of returned confidence interval, default = 0.95.}
}
\value{
Concentration with confidence interval.
}
\description{
Function to take a random sample of a given size from of a given universe.
A random statistical sample of a created universe is taken and converted to concentration result.
}
\examples{
# Create a default universe simulating 50 ng DNA input.
universe_50 = universe(50)

# Take a sample of 20,000 droplets from this universe.
sample = sample_from_universe(universe_50, 20000)
}
