% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ratios.R
\name{simulate_ratios}
\alias{simulate_ratios}
\title{Simulate ratios to be taken from two existing universes.}
\usage{
simulate_ratios(universe_1, universe_2, n_droplets = 20000,
  n_simulations = 100, alpha = 0.95)
}
\arguments{
\item{universe_1}{A universe.}

\item{universe_2}{A universe.}

\item{n_droplets}{n}

\item{n_simulations}{n}

\item{alpha}{Level of significance, default = 0.95.}
}
\value{
Simulation results.
}
\description{
As described in Dube et al., the geometric interpretation of Fieller's theorem can be used to calculate the confidence intervals for a ratio of two values with known confidence intervals.
}
\examples{
#
}
