% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sample.R
\name{simulate_sample}
\alias{simulate_sample}
\title{Simulate a sample: return a random selection of droplets.}
\usage{
simulate_sample(sample_size = 20000, universe = NULL)
}
\arguments{
\item{sample_size}{A table with the results of the simulation.}

\item{universe}{The true value which was simulated.}
}
\value{
List with statistics.
}
\description{
Lorem.
}
\examples{
# Create two default universes simulating 50 ng DNA input.
}
