% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats}
\alias{stats}
\title{Calculate statistics from simulation result table.}
\usage{
stats(results, true_value)
}
\arguments{
\item{results}{A table with the results of the simulation.}

\item{true_value}{The true value which was simulated.}
}
\value{
List with statistics.
}
\description{
Lorem.
}
\examples{
# Create two default universes simulating 50 ng DNA input.
universe_target = universe(50)
universe_reference = universe(50)

# Take two samples
sample = sample(1:length(universe_target), 20000)
conc_target = sample_from_universe(universe_target, sample)
conc_reference = sample_from_universe(universe_target, sample)

# Calculate the ratio
ratio = calc_ratio(conc_target, conc_reference)
ratio
}
